Enable OSGi compliance on package layer via 'DynamicImport-Package: *' instead of Declarative Service (DS) approach. Restructuring and cleanup of initialization process to ensure that all internal config files are found by the corresponding bundle classloaders.
SMACK-343
