+++ smack-android/src/main/java/org/jivesoftware/smack/android/AndroidSmackInitializer.java
@@ -20,13 +20,13 @@ import java.util.List;
 
 import org.apache.http.conn.ssl.StrictHostnameVerifier;
 import org.jivesoftware.smack.SmackConfiguration;
-import org.jivesoftware.smack.initializer.SimpleSmackInitializer;
+import org.jivesoftware.smack.initializer.SmackInitializer;
 import org.jivesoftware.smack.util.stringencoder.Base64;
 import org.jivesoftware.smack.util.stringencoder.Base64UrlSafeEncoder;
 import org.jivesoftware.smack.util.stringencoder.android.AndroidBase64Encoder;
 import org.jivesoftware.smack.util.stringencoder.android.AndroidBase64UrlSafeEncoder;
 
-public class AndroidSmackInitializer extends SimpleSmackInitializer {
+public class AndroidSmackInitializer implements SmackInitializer {
 
 	@Override
 	public List<Exception> initialize() {
