+++ smack-core/build.gradle
@@ -1,3 +1,8 @@
+// Note that this is also declared in the main build.gradle for
+// subprojects, but since evaluationDependsOnChildren is enabled we
+// need to declare it here too
+apply plugin: 'osgi'
+
 description = """\
 Smack core components."""
 
@@ -32,3 +37,9 @@ task createVersionResource(type: CreateFileTask) {
 }
 
 compileJava.dependsOn(createVersionResource)
+
+jar {
+  manifest {
+    instruction 'DynamicImport-Package', '*'
+  }
+}
