+++ smack-core/src/main/java/org/jivesoftware/smack/initializer/SmackInitializer.java
@@ -30,6 +30,7 @@ import org.jivesoftware.smack.SmackConfiguration;
  *
  */
 public interface SmackInitializer {
+	
     public List<Exception> initialize();
-    public List<Exception> initialize(ClassLoader classLoader);
+    
 }
