+++ smack-core/src/main/resources/org.jivesoftware.smack/smack-config.xml
@@ -11,10 +11,10 @@
         <className>org.jivesoftware.smack.util.dns.javax.JavaxResolver</className>
         <className>org.jivesoftware.smack.util.dns.minidns.MiniDnsResolver</className>
         <className>org.jivesoftware.smack.util.dns.dnsjava.DNSJavaResolver</className>
-        <className>org.jivesoftware.smack.initializer.extensions.ExtensionsInitializer</className>
-        <className>org.jivesoftware.smack.initializer.experimental.ExperimentalInitializer</className>
-        <className>org.jivesoftware.smack.initializer.legacy.LegacyInitializer</className>
-        <className>org.jivesoftware.smack.initializer.tcp.SmackTcpSmackInitializer</className>
+        <className>org.jivesoftware.smack.extensions.ExtensionsInitializer</className>
+        <className>org.jivesoftware.smack.experimental.ExperimentalInitializer</className>
+        <className>org.jivesoftware.smack.legacy.LegacyInitializer</className>
+        <className>org.jivesoftware.smack.tcp.TCPInitializer</className>
         <className>org.jivesoftware.smack.sasl.javax.SASLJavaXSmackInitializer</className>
         <className>org.jivesoftware.smack.sasl.provided.SASLProvidedSmackInitializer</className>
         <className>org.jivesoftware.smack.android.AndroidSmackInitializer</className>
