--- smack-experimental/src/main/java/org/jivesoftware/smack/initializer/experimental/ExperimentalInitializer.java
@@ -1,37 +1,0 @@
-/**
- *
- * Copyright 2014 Florian Schmaus
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.jivesoftware.smack.initializer.experimental;
-
-import org.jivesoftware.smack.initializer.UrlInitializer;
-
-/**
- * Initializes the providers in the experimental code stream.
- * 
- * @author Florian Schmaus
- */
-public class ExperimentalInitializer extends UrlInitializer {
-
-    @Override
-    protected String getProvidersUrl() {
-        return "classpath:org.jivesoftware.smackx/experimental.providers";
-    }
-
-    @Override
-    protected String getConfigUrl() {
-        return "classpath:org.jivesoftware.smackx/experimental.xml";
-    }
-}
