+++ smack-experimental/src/main/resources/org.jivesoftware.smack.experimental/experimental.providers
