+++ smack-experimental/src/test/java/org/jivesoftware/smackx/ExperimentalInitializerTest.java
@@ -20,7 +20,7 @@ import static org.junit.Assert.assertTrue;
 
 import java.util.List;
 
-import org.jivesoftware.smack.initializer.experimental.ExperimentalInitializer;
+import org.jivesoftware.smack.experimental.ExperimentalInitializer;
 import org.junit.Test;
 
 public class ExperimentalInitializerTest {
