+++ smack-extensions/src/main/resources/org.jivesoftware.smack.extensions/extensions.providers
