+++ smack-extensions/src/test/java/org/jivesoftware/smackx/ExtensionsInitializerTest.java
@@ -20,7 +20,7 @@ import static org.junit.Assert.assertTrue;
 
 import java.util.List;
 
-import org.jivesoftware.smack.initializer.extensions.ExtensionsInitializer;
+import org.jivesoftware.smack.extensions.ExtensionsInitializer;
 import org.junit.Test;
 
 public class ExtensionsInitializerTest {
