--- smack-java7/src/main/resources/org.jivesoftware.smack/smack-java7-components.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.2.0"
-	enabled="true" immediate="true" name="Smack Java7 API">
-	<implementation
-		class="org.jivesoftware.smack.java7.Java7SmackInitializer" />
-</scr:component>
