+++ smack-legacy/src/main/resources/org.jivesoftware.smack.legacy/workgroup.providers
