--- smack-legacy/src/main/resources/org.jivesoftware.smackx/smack-legacy-components.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.2.0"
-	enabled="true" immediate="true" name="Smack Legacy API">
-	<implementation
-		class="org.jivesoftware.smack.initializer.legacy.LegacyInitializer" />
-</scr:component>
