+++ smack-legacy/src/test/java/org/jivesoftware/smackx/LegacyInitializerTest.java
@@ -20,7 +20,7 @@ import static org.junit.Assert.assertTrue;
 
 import java.util.List;
 
-import org.jivesoftware.smack.initializer.legacy.LegacyInitializer;
+import org.jivesoftware.smack.legacy.LegacyInitializer;
 import org.junit.Test;
 
 public class LegacyInitializerTest {
