--- smack-sasl-provided/src/main/resources/org.jivesoftware.smack/smack-sasl-provided-components.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.2.0"
-	enabled="true" immediate="true" name="Smack Resolver JavaX API">
-	<implementation
-		class="org.jivesoftware.smack.sasl.provided.SASLProvidedSmackInitializer" />
-</scr:component>
