+++ smack-tcp/src/main/resources/org.jivesoftware.smack.tcp/smacktcp.providers
